+++ udata/__init__.py
@@ -5,5 +5,5 @@ uData
 '''
 from __future__ import unicode_literals
 
-__version__ = '1.6.5'
+__version__ = '1.6.6.dev'
 __description__ = 'Open data portal'