+++ udata_gouvfr/__init__.py
@@ -4,5 +4,5 @@ uData customizations for Data.gouv.fr
 '''
 from __future__ import unicode_literals
 
-__version__ = '1.6.3.dev'
+__version__ = '1.6.3'
 __description__ = 'uData customizations for Data.gouv.fr'